requirements.txt
requests
pyyaml
unittest
json